requirements.txt
alembic==1.7.5
anyio==3.3.4
asgiref==3.4.1
backcall==0.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
debugpy==1.5.1
decorator==5.1.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
entrypoints==0.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
ipykernel==6.5.0
ipython==7.29.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
jupyter-client==7.0.6
jupyter-core==4.9.1
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
orjson==3.6.4
parso==0.8.2
passlib==1.7.4
pickleshare==0.7.5
prompt-toolkit==3.0.22
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
pywin32==302
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
